no longer using one-hot

+++ dcq/algorithm.h
@@ -19,10 +19,10 @@ namespace dcq::algorithm {
 
     void compute_colors(dcq::Parameters &params, torch::Tensor &b, torch::Tensor &a);
 
-    bool update_M(int iy, int ix, float *M_data, float *Y_data, float *p_data, float *bii_data,
+    bool update_M(int iy, int ix, int *M_data, float *Y_data, float *p_data, float *bii_data,
                   int ks, int h, int w, int c, int K);
 
-    void update_p(int iy, int ix, float *M_data, float *Y_data, float *a_data, float *b0_data, float *p_data,
+    void update_p(int iy, int ix, int *M_data, float *Y_data, float *a_data, float *b0_data, float *p_data,
                   int ks, int h, int w, int c, int K);
 
     torch::Tensor compute_p(dcq::Parameters &params, torch::Tensor &a, torch::Tensor &b);
